refactor: コメントの追記、モックの調整、型の整合性

+++ modules/API/ForceSensorClient.cpp
@@ -1,3 +1,9 @@
+/**
+ * @file   ForceSensorClient.cpp
+ * @brief  フォースセンサー操作のためのクライアントAPIの実装
+ * @author takuchi17
+ */
+
 #include "ForceSensorClient.h"
 
 ForceSensorClient::ForceSensorClient(SpikeClient& client) : spikeClient(client) {}
@@ -9,8 +15,8 @@ bool ForceSensorClient::isPressed(float threshold)
   return res.value_or(false);
 }
 
-int32_t ForceSensorClient::getForce()
+float ForceSensorClient::getForce()
 {
-  auto res = spikeClient.executeQuery<int32_t>(spike::CommandId::FORCE_SENSOR_GET_FORCE);
-  return res.value_or(0);
+  auto res = spikeClient.executeQuery<float>(spike::CommandId::FORCE_SENSOR_GET_FORCE);
+  return res.value_or(0.0f);
 }
