+++ modules/API/ForceSensorClient.h
@@ -1,3 +1,9 @@
+/**
+ * @file   ForceSensorClient.h
+ * @brief  フォースセンサー操作のためのクライアントAPI
+ * @author takuchi17
+ */
+
 #ifndef FORCESENSOR_CLIENT_H
 #define FORCESENSOR_CLIENT_H
 
@@ -6,12 +12,25 @@
 
 class ForceSensorClient {
  public:
+  /**
+   * @brief コンストラクタ
+   * @param client SpikeClientの参照
+   */
   explicit ForceSensorClient(SpikeClient& client);
+  /**
+   * @brief 押されているか
+   * @param threshold 閾値
+   * @return 押されている場合true、それ以外false
+   */
   bool isPressed(float threshold = 0.5f);
-  int32_t getForce();
+  /**
+   * @brief フォース値を取得する
+   * @return フォース値
+   */
+  float getForce();
 
  private:
-  SpikeClient& spikeClient;
+  SpikeClient& spikeClient;  // SpikeClientの参照
 };
 
 #endif
